chore: move inside content folder
