+++ src/content/content-metadata-html.js
